pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "A Discord bot for content processing and audio generation"
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.0",
    "mistralai>=0.0.7",
    "PyPDF2>=3.0.0",
    "httpx>=0.25.0",
    "pydub>=0.25.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
